fixed cookies and local storage to checks local storage in dev and cookies in prod

+++ src/actions/userAuth-actions.js
@@ -7,7 +7,11 @@ export const signIn = token => ({
 });
 
 export const signOut = () => {
-  util.deleteCookie('Bracket-Busters-Token');
+  if(process.env.NODE_ENV === 'production') {
+    util.deleteCookie('Bracket-Busters-Token');
+  } else {
+    delete localStorage.token;
+  }
   return { type: 'SIGN_OUT' };
 };
 
@@ -19,11 +23,15 @@ export const signUpRequest = user => dispatch => {
     .send(user)
     .then( res => {
       dispatch(signIn(res.text));
-      try {
-        // localStorage.token = res.text.token;
-        util.createCookie('Bracket-Busters-Token', res.text.token, 30);
-      } catch (err) {
-        console.error(err);
+      if(process.env.NODE_ENV === 'production') {
+        try {
+          util.createCookie('Bracket-Busters-Token', res.text.token, 30);
+        } catch (err) {
+          console.error(err);
+        }
+      }
+      else {
+        localStorage.token = res.text;
       }
       return res;
     });
@@ -35,11 +43,15 @@ export const signInRequest = user => dispatch => {
     .auth(user.username, user.password)
     .then( res => {
       dispatch(signIn(res.text));
-      try {
-        // localStorage.token = res.text.token;
-        util.createCookie('Bracket-Busters-Token', res.text.token, 30);
-      } catch (err) {
-        console.error(err);
+      if(process.env.NODE_ENV === 'production') {
+        try {
+          util.createCookie('Bracket-Busters-Token', res.text.token, 30);
+        } catch (err) {
+          console.error(err);
+        }
+      }
+      else {
+        localStorage.token = res.text;
       }
       return res;
     });
